+++ src/main/java/edu/uws/ii/springboot/repositories/IUnitsRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IUnitsRepository extends JpaRepository<Unit, Integer>, JpaSpecificationExecutor<Unit> {
+public interface IUnitsRepository extends JpaRepository<Unit, Long>, JpaSpecificationExecutor<Unit> {
 }
